pyproject.toml
[project]
name = "mcp-image-tools"
version = "0.1.0"
description = "MCP server for image processing tools (chromakey, resize, compress, convert)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "tsilva" }]
dependencies = [
    "mcp[cli]>=1.2.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
mcp-image-tools = "mcp_image_tools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_image_tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
